move buffer-helper to utils (is generic, non-hls.js specific)
