+++ src/utils/buffer-helper.js
