+++ tests/unit/utils/buffer-helper.js
